+++ ChangeLog.md
@@ -10,7 +10,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- shell/nice function
+- function `shell/nice`
+- function `index-of` and `last-index-of` for sequences
 
 - :openai module support for assistant api
 - :openai module support for assistant threads api